requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
Authlib==1.3.0
backoff==2.2.1
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
cryptography==42.0.7
datasets==2.14.7
dill==0.3.7
distro==1.9.0
dspy==0.1.5
dspy-ai==2.1.9
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2023.10.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-health-checking==1.64.0
grpcio-tools==1.64.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
multiprocess==0.70.15
numpy==1.26.4
openai==1.30.3
optuna==3.4.0
packaging==24.0
pandas==2.1.4
protobuf==5.27.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.8.0
urllib3==2.2.1
validators==0.28.1
weaviate-client==3.26.2
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
